# academy/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Summer Academy educational platform",
        "title": "Summer Academy API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/admin/": {
            "get": {
                "security": [
                    {
                        "JWTCookie": []
                    }
                ],
                "description": "Displays admin dashboard with platform statistics",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Admin dashboard",
                "responses": {
                    "200": {
                        "description": "Admin dashboard"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    }
                }
            }
        },
...
        "/admin/users": {
            "get": {
                "security": [
                    {
                        "JWTCookie": []
                    }
                ],
                "description": "Displays a list of all registered users for admin management",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List all users",
                "responses": {
                    "200": {
                        "description": "Users list"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/profile": {
            "get": {
                "security": [
                    {
                        "JWTCookie": []
                    }
                ],
                "description": "Displays user profile with submission history and statistics",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Show user profile",
                "responses": {
                    "200": {
                        "description": "Profile page"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWTCookie": []
                    }
                ],
                "description": "Update user profile information like username",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Update user profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User's username",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWT": {
            "description": "Bearer JWT token for authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "JWTCookie": {
            "description": "JWT token stored in cookie for authentication",
            "type": "apiKey",
            "name": "academy_session",
            "in": "cookie"
        }
    }
}